requirements.txt
streamlit
openai
requests
langchain
faiss-cpu
datasets
python-dotenv
tiktoken